requirements.txt
streamlit
absl-py==0.12.0
altair==4.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
audioread==2.1.9
Automat==20.2.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
blis==0.7.4
bokeh==2.3.2
breadability==0.1.20
cachetools==4.2.1
catalogue==2.0.6
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
clang==5.0
click==7.1.2
colorama==0.4.4
configparser==5.0.2
constantly==15.1.0
cryptography==3.4.6
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.5
dalex==1.2.0
darr==0.3.1
datasets==1.6.2
DateTime==4.3
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.3
distlib==0.3.1
docker-pycreds==0.4.0
docopt==0.6.2
eli5==0.11.0
emnist==0.0
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
flatbuffers==1.12
fpdf==1.7.2
fsspec==2021.4.0
future==0.18.2
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
google-trans-new==1.1.9
googleapis-common-protos==1.53.0
googletrans==3.0.0
graphviz==0.16
greenlet==1.1.0
grpcio==1.41.0
gTTS==2.2.2
h11==0.9.0
h2==3.2.0
h5py==3.1.0
helper==2.5.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.0.8
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
imageio==2.9.0
incremental==17.5.0
iniconfig==1.1.1
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywebrtc==0.6.0
ipywidgets==7.6.3
itemadapter==0.2.0
itemloaders==1.0.4
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.6.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
langdetect==1.0.7
librosa==0.8.0
lime==0.2.0.1
llvmlite==0.36.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
mistune==0.8.4
multiprocess==0.70.11.1
multitasking==0.0.9
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
neattext==0.1.0
nest-asyncio==1.5.1
networkx==2.5.1
nlp==0.4.0
nltk==3.6.3
notebook==6.4.4
numba==0.53.1
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.2.54
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandasql==0.7.3
pandocfilters==1.4.3
parsel==1.6.0
parso==0.8.2
pathtools==0.1.2
pathy==0.5.2
patsy==0.5.1
pickleshare==0.7.5
Pillow==8.3.2
pipreqs==0.4.10
plac==1.1.3
plotly==4.14.3
plotly-express==0.4.1
pluggy==0.13.1
pooch==1.3.0
preshed==3.0.5
priority==1.3.0
prometheus-client==0.10.1
promise==2.3
prompt-toolkit==3.0.18
Protego==0.1.16
protobuf==3.15.8
psutil==5.8.0
py==1.10.0
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybase64==1.1.4
pycountry==20.7.3
pycparser==2.20
pydantic==1.7.4
pydeck==0.6.2
PyDispatcher==2.0.5
pydub==0.25.1
Pygments==2.8.1
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
queuelib==1.5.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
retrying==1.3.3
rfc3986==1.5.0
rsa==4.7.2
sacremoses==0.0.45
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.2
Scrapy==2.5.1
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
sentencepiece==0.1.95
sentry-sdk==1.1.0
seqeval==1.2.2
service-identity==18.1.0
shortuuid==1.0.1
simpletransformers==0.61.4
six==1.15.0
sklearn==0.0
smart-open==3.0.0
smmap==4.0.0
sniffio==1.2.0
sound==0.1.0
sounddevice==0.4.1
SoundFile==0.10.3.post1
speech==0.5.2
SpeechRecognition==3.8.1
SQLAlchemy==1.4.14
srsly==2.4.1
statsmodels==0.12.2
streamlit==1.0.0
streamlit-bokeh-events==0.1.2
subprocess32==3.5.4
sumy==0.8.1
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
tensorflow==2.6.0
tensorflow-datasets==4.2.0
tensorflow-estimator==2.6.0
tensorflow-metadata==0.29.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
textblob==0.15.3
thinc==8.0.10
threadpoolctl==2.1.0
tifffile==2021.6.14
time-str==0.0.3.post1
tokenizers==0.10.2
toml==0.10.2
toolz==0.11.1
torch==1.8.1
tornado==6.1
tqdm==4.49.0
traitlets==5.0.5
transformers==4.5.1
typer==0.3.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.7
validators==0.18.2
virtualenv==20.4.3
w3lib==1.22.0
wandb==0.10.30
wasabi==0.8.2
watchdog==2.0.2
wavio==0.0.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==2.0.1
xxhash==2.0.2
yarg==0.1.9
yfinance==0.1.59
zope.interface==5.2.0
opencv-python
deepface
